+++ src/main/java/edu/usfca/cs/mr/GeoSnowDepthGT0/GeoSnowDepth0Mapper.java
@@ -10,7 +10,7 @@ import java.util.StringTokenizer;
 /**
  * Created by xuekang on 10/27/17.
  */
-public class GeoSnowDepth0Mapper extends Mapper<LongWritable, Text, Text, DoubleWritable> {
+public class GeoSnowDepth0Mapper extends Mapper<LongWritable, Text, Text, FloatWritable> {
     @Override
     protected void map(LongWritable key, Text value, Context context)
             throws IOException, InterruptedException {
@@ -30,6 +30,6 @@ public class GeoSnowDepth0Mapper extends Mapper<LongWritable, Text, Text, Double
 //        else{
 //            context.write(new Text(Geohash), new BooleanWritable(false));
 //        }
-        context.write(new Text(Geohash), new DoubleWritable(snowDepth));
+        context.write(new Text(Geohash), new FloatWritable(snowDepth));
     }
 }
